requirements.txt
requests
xmljson
python-slugify